refactor: rename furniture event packages in NexoAddon
