+++ src/main/java/zone/vao/nexoAddon/events/nexo/furnitures/NexoFurnitureBreakListener.java
@@ -4,8 +4,8 @@ import com.nexomc.nexo.api.events.furniture.NexoFurnitureBreakEvent;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.EventPriority;
 import org.bukkit.event.Listener;
-import zone.vao.nexoAddon.events.nexo.furnitures.furnitureBreaks.BreakDoubleHit;
-import zone.vao.nexoAddon.events.nexo.furnitures.furnitureBreaks.BreakWithToolsListener;
+import zone.vao.nexoAddon.events.nexo.furnitures.breaks.BreakDoubleHit;
+import zone.vao.nexoAddon.events.nexo.furnitures.breaks.BreakWithToolsListener;
 
 
 public class NexoFurnitureBreakListener implements Listener {
