+++ src/main/java/zone/vao/nexoAddon/events/nexo/furnitures/NexoFurnitureInteractListener.java
@@ -4,8 +4,8 @@ import com.nexomc.nexo.api.events.furniture.NexoFurnitureInteractEvent;
 import org.bukkit.entity.Player;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.Listener;
-import zone.vao.nexoAddon.events.nexo.furnitures.playerFurnitureInteracts.DisplayCropsHologram;
-import zone.vao.nexoAddon.events.nexo.furnitures.playerFurnitureInteracts.Fertilize;
+import zone.vao.nexoAddon.events.nexo.furnitures.interacts.DisplayCropsHologram;
+import zone.vao.nexoAddon.events.nexo.furnitures.interacts.Fertilize;
 
 import java.util.UUID;
 import java.util.concurrent.ConcurrentHashMap;
